pyproject.toml
[tool.poetry]
name = "ai-agent-demos"
version = "0.1.0"
description = "Enterprise AI Agent Platform with LangGraph and Databricks"
authors = ["Your Team <team@example.com>"]
readme = "README.md"
packages = [
    { include = "agents" },
    { include = "shared" },
]

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.2.0"
langchain = "^0.2.0"
langchain-openai = "^0.1.0"
langchain-community = "^0.2.0"
pydantic = "^2.7.0"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
databricks-sdk = "^0.28.0"
mlflow = "^2.14.0"
opentelemetry-api = "^1.25.0"
opentelemetry-sdk = "^1.25.0"
opentelemetry-instrumentation-fastapi = "^0.46b0"
pymsteams = "^0.2.2"
python-multipart = "^0.0.9"
pyyaml = "^6.0.1"
pypdf = "^4.2.0"
chromadb = "^0.5.0"
tiktoken = "^0.7.0"
reportlab = "^4.2.0"
pandas = "^2.2.0"
numpy = "^1.26.0"
scikit-learn = "^1.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.23.0"
black = "^24.4.0"
ruff = "^0.4.0"
mypy = "^1.10.0"
pre-commit = "^3.7.0"
httpx = "^0.27.0"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --cov=agents --cov=shared --cov-report=html --cov-report=term-missing --cov-fail-under=80"
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
